+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [unreleased]
+### Added
+- Parsing of `variantType`, `referenceCopyNumber`, `copyNumber` and `chromosomeCoordinates` fields from CSV/TSV files
+- Genome assembly can be passed as an optional param to tsv_2_json and csv_to_json to create submission containing chromosome coordinates
+
 ## [2.0.1]
 ### Fixed
 - Parsing of `Condition ID type`